revamp of how I sort the analyzing scripts. Step 1: Set the rootfile global variable as the output file. Step 2: Execute RN_RootInit() Step 3: create instances of your custom RN_Analyzer class. Step 4: Add them to the analyzers TList. Step 5: Init the MainAnalyzer class with the input root file. 	Also add as many more files you want to sort by doing 	MainAnalyzer.AddTree() Step 6: Load Variable File(s) then execute SetCalibrations(). Step 7: Load Gate File(s) so that your analyzer(s) can use them. Step 8: MainAnalyzer.Loop();
This will run over all of the root files with GetEntry.  RN_Analyzer
is pre-set to SetBranchAddress for all the electronics modules. If
more modules are needed be sure to add them to the unpack2root as well as
the rn_analyzer.  For each GetEntry, GetDetectorEntry is also executed which
serves as the channelmap. for instance: RN_RFtime.fT = TDC1[0];