WARNING: Kevin, if you pull this make sure to move your modified analyzers to another folder so they don't get overwritten.  We need to merge the changes we made at some point.
Major Update:
Comments added to some classes.
RN_Analyzer expanded to have a Process() and ProcessFill() to seperate the functions of doing calculations/gate checks from doing histogram filling.
this allows the user to set filtering requirements in Process() and by returning false in any Process() function (that was added to the analyzer list) then NO
ProcessFills() will take place.

RN_Analyzer also now takes the "Event" branch from the unpack2root to look at RunNum and (for the future) scalerIDX.

S2_Neut_IC_Analyzer was changed to S2_Neut_IC_NaI_Analyzer and this change is reflected in the linkDef and Makefile

An Analyzer_Template was added as a documentation/tutorial to adding simple analyzers.

RN_Root global variable E_fragment added for the purpose of Q_value reconstruction.  set it in calibration file.