RN_EventProcessor member names changed for consistency. The output TFile is now a member of the RN_EventProcessor and is a static global member of RN_EventProcessor. This was done to continue to simplify RN_Root.cpp which is intended to just setup the modules/detectors for the experiment and the methods to handle them.
All analyzers edited to reflect this change.

ScalerValues are now a fixed array of 32 channels (as is the buffer format for a ScalerBuffer which always reads out 32 channels)
This means that InitStack is now deprecated.  It is the user's responsibility to keep track of what the inputs were for each scaler.

RNBatchMode was changed to reflect this.  Executing RNBatchMode now requires just two inputs:
RNBatchMode <filename.evt> <output.root>
or
RNBatchMode <filedirectory> <output.root>

The scaler sums and averages are now dumped into the log file