RNFilters have the filters used during the F17(dn) experiment June 2014 renamed some of the member variables to follow the root/taligent conventions
RNBaseDetector:
added the elin and eshift global calibration variables.  Broke up the Energy
and time calibration functions E() into:
E_Offset(i) =  (fE[i] + offset[chan])
E_Local(i) = ( E_Offset(i) * a1(chan) + a0(chan))
E(i) = ( E_local(i) * elin + eshift )
ERaw(i) = fE[i]

The parameters stored in the BaseDetector class are now protected and should
only be changed when the InsertHit or Reset function is called.
The raw data is accessed using functions ERaw and TRaw instead of directly.

*Major Change:
InsertHit algorithm will continue REGARDLESS of the energy value
IF there is a timing signal.
Also, a check if fMult is already as high as fNumOfCh, if it is a warning is given and an option to abort the program.
this option is provided because fMult should never even attempt to incremement as many times as to fill all of the fNumOfChs if the channel map is set
properly.

RN_Analyzer is altered to reflect this change.